+++ src/mocks/index.js
@@ -1,5 +1,5 @@
 /*
- * Copyright 2021 Adobe. All rights reserved.
+ * Copyright 2023 Adobe. All rights reserved.
  * This file is licensed to you under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License. You may obtain a copy
  * of the License at http://www.apache.org/licenses/LICENSE-2.0
@@ -10,15 +10,6 @@
  * governing permissions and limitations under the License.
  */
 
-export * from './blob-storage.js';
-export * as commonTypes from './common-typedefs.js';
-export * as contextHelper from './context.js'; // preserve backwards compatibility
-export { ContextHelper } from './context.js';
-export * from './fetch-retry.js';
+export * from './local-search-index-storage.js';
 export * from './queue.js';
-export * from './rest-error.js';
-export * from './router.js';
-export * from './schema-validator.js';
-export * from './secret.js';
-export { CloudSearchIndexStorage } from './cloud-search-index-storage.js';
-export * as mocks from './mocks/index.js';
+export * from './secrets.js';