+++ CHANGELOG.md
@@ -1,3 +1,24 @@
+## [1.6.1](https://github.com/adobe/content-lake-commons/compare/v1.6.0...v1.6.1) (2023-05-25)
+
+
+### Bug Fixes
+
+* adding delete to the search mock ([#29](https://github.com/adobe/content-lake-commons/issues/29)) ([72a55b8](https://github.com/adobe/content-lake-commons/commit/72a55b85048f15183908ed141660725371092d56))
+
+# [1.6.0](https://github.com/adobe/content-lake-commons/compare/v1.5.1...v1.6.0) (2023-05-23)
+
+
+### Features
+
+* adding support to index storage to search across arbitrary fields ([c5c8421](https://github.com/adobe/content-lake-commons/commit/c5c84217cd3a9c65ade6b59c272ac308e0c69a44))
+
+## [1.5.1](https://github.com/adobe/content-lake-commons/compare/v1.5.0...v1.5.1) (2023-05-20)
+
+
+### Bug Fixes
+
+* **deps:** update external fixes ([2d9a4e9](https://github.com/adobe/content-lake-commons/commit/2d9a4e976eaed28f8442d3ca5879b58a05649f47))
+
 # [1.5.0](https://github.com/adobe/content-lake-commons/compare/v1.4.0...v1.5.0) (2023-05-12)